RAG_iti/requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiohttp==3.8.3
aiosignal==1.3.1
array-record==0.2.0
astunparse==1.6.3
async-timeout==4.0.2
bert-score==0.3.13
cachetools==5.3.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dacite==1.8.1
datasets==2.3.0
dill==0.3.5.1
dm-tree==0.1.8
einops==0.6.1
et-xmlfile==1.1.0
etils==1.3.0
fairscale==0.4.13
fancy-einsum==0.0.3
filelock==3.8.0
fire==0.5.0
flagembedding==1.2.11
flatbuffers==23.5.8
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2024.3.1
future==0.18.3
gast==0.4.0
geotorch==0.3.0
gin-config==0.5.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
h5py==3.7.0
hickle==5.0.2
huggingface-hub==0.26.3
ipdb==0.13.9
jax==0.4.9
jieba==0.42.1
joblib==1.2.0
keras==2.12.0
keyboard==0.13.5
kiwisolver==1.4.4
libclang==16.0.0
llvmlite==0.39.1
markdown==3.4.3
matplotlib==3.7.1
mesh-tensorflow==0.1.21
ml-dtypes==0.1.0
multidict==6.0.3
multiprocess==0.70.13
nltk==3.8.1
numba==0.56.4
nvidia-ml-py==12.535.133
nvitop==1.3.2
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
opt-einsum==3.3.0
pandas==2.0.1
pandas-stubs==1.5.1.221024
parallelformers==1.2.7
peft==0.12.0
plotly==5.14.1
portalocker==2.7.0
promise==2.3
protobuf==4.23.0
pyarrow==17.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pynndescent==0.5.8
pyyaml==6.0
regex==2022.10.31
requests==2.32.3
requests-oauthlib==1.3.1
responses==0.18.0
rouge-score==0.1.2
rsa==4.9
sacrebleu==2.3.1
sacremoses==0.0.53
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentence-transformers==3.1.0
t5==0.7.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
tensorflow-text==2.12.1
termcolor==2.2.0
tf-slim==1.1.0
tfds-nightly==4.9.2.dev202305230044
threadpoolctl==3.1.0
tokenizers==0.19.1
toml==0.10.2
torchmetrics==0.6.2
tqdm==4.67.1
transformers==4.40.0
types-pytz==2022.6.0.1
tzdata==2023.3
umap-learn==0.5.3
werkzeug==2.3.4
wrapt==1.14.1
xxhash==3.1.0
yarl==1.8.2
pyvene==0.2.2